// apps/admin/web/src/types/garage.ts
export interface ClusterSummary {
  id: string;
  name: string;
  endpoint: string;
  createdAt: string;
  updatedAt?: string;
}
// ...
export interface FreeSpaceResp {
  total: number;
  available: number;
}

export interface NodeAssignedRole {
  zone: string;
  tags: string[];
  capacity?: number | null;
}

export interface NodeResp {
  id: string;
  isUp: boolean;
  draining: boolean;
  addr?: string | null;
  hostname?: string | null;
  garageVersion?: string | null;
  lastSeenSecsAgo?: number | null;
  role?: NodeAssignedRole | null;
  dataPartition?: FreeSpaceResp | null;
  metadataPartition?: FreeSpaceResp | null;
}

export interface NodeInfoResponse {
  nodeId: string;
  garageVersion: string;
  rustVersion: string;
  dbEngine: string;
  garageFeatures?: string[] | null;
}

export interface GetClusterStatusResponse {
  layoutVersion: number;
  nodes: NodeResp[];
}

export interface GetClusterHealthResponse {
  status: string;
  knownNodes: number;
  connectedNodes: number;
  storageNodes: number;
  storageNodesUp: number;
  partitions: number;
  partitionsQuorum: number;
  partitionsAllOk: number;
}